requirements.txt
Flask-3.1.0
pyodbc-5.2.0